:sparkles: Server permissions added

+++ src/server/src/controllers/projects/projects.controller.ts
@@ -1,8 +1,8 @@
 /******************************************************************************
- * @Author                : Jbristhuille<jean-baptiste@halfsquare.fr>         *
+ * @Author                : 0K00<qdouvillez@gmail.com>                        *
  * @CreatedDate           : 2023-06-13 14:10:50                               *
- * @LastEditors           : Jbristhuille<jean-baptiste@halfsquare.fr>         *
- * @LastEditDate          : 2023-11-16 17:29:36                               *
+ * @LastEditors           : 0K00<qdouvillez@gmail.com>                        *
+ * @LastEditDate          : 2024-01-17 14:39:22                               *
  *****************************************************************************/
 
 /* SUMMARY
@@ -30,7 +30,8 @@ import {
   HttpException,
   HttpStatus,
   UseGuards,
-  UseInterceptors
+  UseInterceptors,
+  SetMetadata
 } from "@nestjs/common";
 import { Project } from "@prisma/client";
 import * as jwt from "jsonwebtoken";
@@ -43,10 +44,12 @@ import * as projectsDto from "./projects.dto";
 /* Services */
 import { PrismaService } from "src/prisma.service";
 import { SocketService } from "src/services/socket/socket.service";
+import { PermsService } from "src/services/perms/perms.service";
 /***/
 
 /* Guards */
 import { IsConnectedGuard } from "src/guards/is-connected.guard";
+import { IsPermissionsGuard } from "src/guards/is-perms.guard";
 /***/
 
 /* Interceptors */
@@ -57,7 +60,8 @@ import { ActivitiesInterceptor } from "../activities/activities.interceptor";
 @UseGuards(IsConnectedGuard)
 export class ProjectsController {
   constructor(private prisma: PrismaService,
-              private socket: SocketService) {
+              private socket: SocketService,
+              private perm: PermsService) {
   }
 
   /**
@@ -65,6 +69,8 @@ export class ProjectsController {
   * @return - Projects list 
   */
   @Get()
+  @UseGuards(IsPermissionsGuard)
+  @SetMetadata("permissions", [{type: "projects", actions: ["view"]}])
   async getAll(): Promise<Project[]> {
     try {
       return await this.prisma.project.findMany({
@@ -85,6 +91,8 @@ export class ProjectsController {
   * @return - Project data 
   */
   @Get(":id")
+  @UseGuards(IsPermissionsGuard)
+  @SetMetadata("permissions", [{type: "projects", actions: ["view"]}])
   async getById(@Param("id") id: string): Promise<Project> {
     try {
       const ret = await this.prisma.project.findUnique({
@@ -107,6 +115,8 @@ export class ProjectsController {
   */
   @Post()
   @UseInterceptors(ActivitiesInterceptor)
+  @UseGuards(IsPermissionsGuard)
+  @SetMetadata("permissions", [{type: "projects", actions: ["create"]}])
   async create(
     @Body() body: projectsDto.CreateInput,
     @Headers("x-token") token: string): Promise<Project> {
@@ -145,33 +155,52 @@ export class ProjectsController {
   @UseInterceptors(ActivitiesInterceptor)
   async update(
     @Param("id") id: string,
-    @Body() body: projectsDto.UpdateInput): Promise<Project> {
+    @Body() body: projectsDto.UpdateInput,
+    @Headers("x-token") token: string): Promise<Project | { perm: boolean; message: string; }> {
     try {
-      const project = await this.prisma.project.update({
-        where: {id: id},
-        data: {
-          ...body,
-          assignments: body.assignments ? {
-            deleteMany: {},
-            create: body.assignments.map((el) => {
-              return {userId: el.userId};
-            })
-          } : undefined,
-          labels: body.labels ? {
-            deleteMany: {},
-            create: body.labels.map((el) => {
-              return {labelId: el.labelId};
-            })
-          } : undefined
-        },
+      const user = await this.prisma.user.findFirst({
+        where: { id: jwt.verify(token, process.env.SALT).id },
         include: {
-          labels: true,
-          assignments: true
+          role: true
         }
       });
 
-      this.socket.broadcast("projects", project);
-      return project;
+      const canUpdate = await this.perm.updateData(
+        body, 
+        id, 
+        "projects", 
+        user, 
+        ["assignments", "actualVersion", "status", "labels", "level", "title", "description"]);
+      
+      if(canUpdate) {
+        const project = await this.prisma.project.update({
+          where: {id: id},
+          data: {
+            ...body,
+            assignments: body.assignments ? {
+              deleteMany: {},
+              create: body.assignments.map((el) => {
+                return {userId: el.userId};
+              })
+            } : undefined,
+            labels: body.labels ? {
+              deleteMany: {},
+              create: body.labels.map((el) => {
+                return {labelId: el.labelId};
+              })
+            } : undefined
+          },
+          include: {
+            labels: true,
+            assignments: true
+          }
+        });
+  
+        this.socket.broadcast("projects", project);
+        return project;
+      } else {
+        throw new HttpException("You do not have the necessary permission to perform this action", HttpStatus.NOT_MODIFIED);
+      }
     } catch (err) {
       throw err;
     }
@@ -185,6 +214,8 @@ export class ProjectsController {
   */
   @Delete(":id")
   @UseInterceptors(ActivitiesInterceptor)
+  @UseGuards(IsPermissionsGuard)
+  @SetMetadata("permissions", [{type: "projects", actions: ["delete"]}])
   async deleteById(@Param("id") id: string): Promise<{message: string}> {
     try {
       await this.prisma.project.delete({where: {id: id}});
