+++ src/server/src/controllers/tasks/tasks.controller.ts
@@ -1,8 +1,8 @@
 /******************************************************************************
- * @Author                : Jbristhuille<jean-baptiste@halfsquare.fr>         *
+ * @Author                : 0K00<qdouvillez@gmail.com>                        *
  * @CreatedDate           : 2023-06-24 13:47:35                               *
- * @LastEditors           : Jbristhuille<jean-baptiste@halfsquare.fr>         *
- * @LastEditDate          : 2023-09-27 14:31:16                               *
+ * @LastEditors           : 0K00<qdouvillez@gmail.com>                        *
+ * @LastEditDate          : 2024-01-17 14:40:04                               *
  *****************************************************************************/
 
 /* SUMMARY
@@ -31,7 +31,8 @@ import {
   HttpStatus,
   Param,
   UseGuards,
-  UseInterceptors
+  UseInterceptors,
+  SetMetadata
 } from "@nestjs/common";
 import { Task } from "@prisma/client";
 import * as jwt from "jsonwebtoken";
@@ -44,10 +45,12 @@ import * as tasksDto from "./tasks.dto";
 /* Services */
 import { SocketService } from "src/services/socket/socket.service";
 import { PrismaService } from "src/prisma.service";
+import { PermsService } from "src/services/perms/perms.service";
 /***/
 
 /* Guards */
 import { IsConnectedGuard } from "src/guards/is-connected.guard";
+import { IsPermissionsGuard } from "src/guards/is-perms.guard";
 /***/
 
 /* Interceptors */
@@ -58,7 +61,8 @@ import { ActivitiesInterceptor } from "../activities/activities.interceptor";
 @UseGuards(IsConnectedGuard)
 export class TasksController {
   constructor(private socket: SocketService,
-              private prisma: PrismaService) {
+              private prisma: PrismaService,
+              private perm: PermsService) {
   }
 
   /**
@@ -66,6 +70,8 @@ export class TasksController {
   * @return - Tasks list
   */
   @Get("tasks")
+  @UseGuards(IsPermissionsGuard)
+  @SetMetadata("permissions", [{type: "tasks", actions: ["view"]}])
   async getAll(): Promise<Task[]> {
     try {
       return await this.prisma.task.findMany({
@@ -86,6 +92,8 @@ export class TasksController {
   * @return - Task's data 
   */
   @Get("tasks/:id")
+  @UseGuards(IsPermissionsGuard)
+  @SetMetadata("permissions", [{type: "tasks", actions: ["view"]}])
   async getById(@Param("id") id: string): Promise<Task> {
     try {
       const task = await this.prisma.task.findUnique({
@@ -110,6 +118,8 @@ export class TasksController {
   * @returns - Task's details
   */
   @Post("projects/:id/tasks")
+  @UseGuards(IsPermissionsGuard)
+  @SetMetadata("permissions", [{type: "tasks", actions: ["create"]}])
   @UseInterceptors(ActivitiesInterceptor)
   async create(
     @Param("id") id: string,
@@ -153,34 +163,53 @@ export class TasksController {
   @UseInterceptors(ActivitiesInterceptor)
   async update(
     @Param("id") id: string,
-    @Body() body: tasksDto.UpdateInput
+    @Body() body: tasksDto.UpdateInput,
+    @Headers("x-token") token: string
   ): Promise<Task> {
     try {
-      const task = await this.prisma.task.update({
-        where: {id: id},
-        data: {
-          ...body,
-          assignments: body.assignments ? {
-            deleteMany: {},
-            create: body.assignments.map((el) => {
-              return {userId: el.userId};
-            })
-          } : undefined,
-          labels: body.labels ? {
-            deleteMany: {},
-            create: body.labels.map((el) => {
-              return {labelId: el.labelId};
-            })
-          } : undefined
-        },
+      const user = await this.prisma.user.findFirst({
+        where: { id: jwt.verify(token, process.env.SALT).id },
         include: {
-          labels: true,
-          assignments: true
+          role: true
         }
       });
 
-      this.socket.broadcast("tasks", task);
-      return task;
+      const canUpdate = await this.perm.updateData(
+        body, 
+        id, 
+        "tasks", 
+        user, 
+        ["assignments", "targetVersionId", "status", "labels", "level", "title", "description"]);
+      
+      if(canUpdate) {
+        const task = await this.prisma.task.update({
+          where: {id: id},
+          data: {
+            ...body,
+            assignments: body.assignments ? {
+              deleteMany: {},
+              create: body.assignments.map((el) => {
+                return {userId: el.userId};
+              })
+            } : undefined,
+            labels: body.labels ? {
+              deleteMany: {},
+              create: body.labels.map((el) => {
+                return {labelId: el.labelId};
+              })
+            } : undefined
+          },
+          include: {
+            labels: true,
+            assignments: true
+          }
+        });
+  
+        this.socket.broadcast("tasks", task);
+        return task;
+      } else {
+        throw new HttpException("You do not have the necessary permission to perform this action", HttpStatus.NOT_MODIFIED);
+      }
     } catch (err) {
       throw err;
     }
@@ -194,6 +223,8 @@ export class TasksController {
   */
   @Delete("tasks/:id")
   @UseInterceptors(ActivitiesInterceptor)
+  @UseGuards(IsPermissionsGuard)
+  @SetMetadata("permissions", [{type: "tasks", actions: ["delete"]}])
   async deleteById(@Param("id") id: string): Promise<{message: string}> {
     try {
       await this.prisma.task.delete({where: {id: id}});
