+++ src/server/src/controllers/users/users.controller.ts
@@ -2,7 +2,7 @@
  * @Author                : 0K00<qdouvillez@gmail.com>                        *
  * @CreatedDate           : 2023-06-01 15:15:39                               *
  * @LastEditors           : 0K00<qdouvillez@gmail.com>                        *
- * @LastEditDate          : 2024-01-16 18:07:46                               *
+ * @LastEditDate          : 2024-01-17 14:40:30                               *
  *****************************************************************************/
 
 /* SUMMARY
@@ -24,6 +24,7 @@ import {
   Get,
   Param,
   Body,
+  Headers,
   HttpException,
   HttpStatus,
   UseGuards,
@@ -33,10 +34,12 @@ import { sha256 } from "js-sha256";
 import * as jwt from "jsonwebtoken";
 import { Assignment } from "@prisma/client";
 import * as fs from "fs";
+import { resolve } from "path";
 /***/
 
 /* Services */
 import { PrismaService } from "src/prisma.service";
+import { PermsService } from "src/services/perms/perms.service";
 /***/
 
 /* Dto */
@@ -46,13 +49,13 @@ import * as usersDto from "./users.dto";
 /* Guards */
 import { IsConnectedGuard } from "src/guards/is-connected.guard";
 import { IsSelfGuard } from "src/guards/is-self.guard";
-import { resolve } from "path";
 import { IsAdminGuard } from "src/guards/is-admin.guard";
 /***/
 
 @Controller("api")
 export class UsersController {
-  constructor(private prisma: PrismaService) {
+  constructor(private prisma: PrismaService, 
+              private perm: PermsService) {
   }
 
   /**
@@ -156,25 +159,51 @@ export class UsersController {
   */
   @Put("users/:id")
   @UseGuards(IsSelfGuard)
-  async updateProfile(@Param("id") id: string, @Body() body: usersDto.UpdateInput): Promise<usersDto.ConnectOutput> {
-    if (body.oldPicture) fs.unlinkSync(resolve("upload")+"/"+body.oldPicture);
+  async updateProfile(
+    @Param("id") id: string, 
+    @Body() body: usersDto.UpdateInput,
+    @Headers("x-token") token: string
+  ): Promise<usersDto.ConnectOutput> {
+    try {
+      const userInfo = await this.prisma.user.findFirst({
+        where: { id: jwt.verify(token, process.env.SALT).id },
+        include: {
+          role: true
+        }
+      });
 
-    delete body.oldPicture;
-
-    let user = await this.prisma.user.update({
-      where: {id: id},
-      include: {
-        role: true
-      },
-      data: body
-    });
-
-    user["token"] = jwt.sign(user, process.env.SALT, {
-      algorithm: "HS256",
-      expiresIn: process.env.SESSION_TIME
-    });
-
-    return new usersDto.ConnectOutput(user);
+      const canUpdate = await this.perm.updateData(
+        body, 
+        id, 
+        "profile", 
+        userInfo, 
+        ["name", "email", "picture"]);
+      
+      if(canUpdate) {
+        if (body.oldPicture) fs.unlinkSync(resolve("upload")+"/"+body.oldPicture);
+  
+        delete body.oldPicture;
+    
+        let user = await this.prisma.user.update({
+          where: {id: id},
+          include: {
+            role: true
+          },
+          data: body
+        });
+    
+        user["token"] = jwt.sign(user, process.env.SALT, {
+          algorithm: "HS256",
+          expiresIn: process.env.SESSION_TIME
+        });
+    
+        return new usersDto.ConnectOutput(user);
+      } else {
+        throw new HttpException("You do not have the necessary permission to perform this action", HttpStatus.NOT_MODIFIED);
+      }
+    } catch (err) {
+      throw err;
+    }
   }
   /***/
